#CTCTOWALTZ-2758
#6644

+++ waltz-service/src/main/java/org/finos/waltz/service/involvement/InvolvementService.java
@@ -19,6 +19,7 @@
 package org.finos.waltz.service.involvement;
 
 import org.finos.waltz.common.Checks;
+import org.finos.waltz.common.SetUtilities;
 import org.finos.waltz.data.EntityReferenceNameResolver;
 import org.finos.waltz.data.GenericSelector;
 import org.finos.waltz.data.GenericSelectorFactory;
@@ -27,7 +28,8 @@ import org.finos.waltz.data.person.PersonDao;
 import org.finos.waltz.model.*;
 import org.finos.waltz.model.changelog.ChangeLog;
 import org.finos.waltz.model.changelog.ImmutableChangeLog;
-import org.finos.waltz.model.involvement.*;
+import org.finos.waltz.model.involvement.EntityInvolvementChangeCommand;
+import org.finos.waltz.model.involvement.Involvement;
 import org.finos.waltz.model.involvement_kind.InvolvementKind;
 import org.finos.waltz.model.person.Person;
 import org.finos.waltz.model.user.SystemRole;
@@ -49,6 +51,7 @@ import static org.finos.waltz.common.FunctionUtilities.time;
 import static org.finos.waltz.common.ListUtilities.applyToFirst;
 import static org.finos.waltz.common.ListUtilities.newArrayList;
 import static org.finos.waltz.common.SetUtilities.map;
+import static org.finos.waltz.common.StringUtilities.isEmpty;
 
 
 @Service
@@ -136,6 +139,7 @@ public class InvolvementService {
                                         EntityInvolvementChangeCommand command) {
 
         checkInvolvementKindIsUserSelectable(command);
+        checkHasEditPermissionForKind(Long.valueOf(command.involvementKindId()), userId);
 
         Involvement involvement = mkInvolvement(entityReference, command);
         boolean result = involvementDao.save(involvement) == 1;
@@ -149,7 +153,10 @@ public class InvolvementService {
     public boolean removeEntityInvolvement(String userId,
                                            EntityReference entityReference,
                                            EntityInvolvementChangeCommand command) {
+
         Involvement involvement = mkInvolvement(entityReference, command);
+        checkHasEditPermissionForKind(Long.valueOf(command.involvementKindId()), userId);
+
         boolean result = involvementDao.remove(involvement) > 0;
         if (result) {
             logChange(entityReference, userId, command);
@@ -296,5 +303,15 @@ public class InvolvementService {
 
     public Set<Involvement> findByKindIdAndEntityKind(long id, EntityKind kind) {
         return involvementDao.findByKindIdAndEntityKind(id, kind);
+    }
+
+
+    private void checkHasEditPermissionForKind(Long kindId, String username) {
+        InvolvementKind kind = involvementKindService.getById(kindId);
+        if (!isEmpty(kind.permittedRole())) {
+            checkTrue(
+                    userRoleService.hasAnyRole(username, SetUtilities.asSet(kind.permittedRole(), SystemRole.ADMIN.name())),
+                    format("User does not have the required permissions to edit involvement kind: '%s'", kind.name()));
+        }
     }
 }
